+++ bin/web-server/routerManager.js
@@ -1,6 +1,7 @@
-export {indexRouter} from '../../routes/index.js'
-export {registerRouter} from '../../routes/register.js';
-export {watchRouter} from '../../routes/watch.js';
-export {downloadRouter} from '../../routes/download.js';
-export {managementRouter, deleteRouter} from '../../routes/management.js';
+export {indexRouter} from '../../routes/index-router.js'
+export {registerRouter} from '../../routes/register-router.js';
+export {watchRouter} from '../../routes/watch-router.js';
+export {downloadRouter} from '../../routes/download-router.js';
+export {managementRouter} from '../../routes/management-router.js';
+export {deleteRouter} from '../../routes/delete-router.js';
 export {loginRouter} from '../../routes/login-router.js';
